+++ spec/Storefront/Controller/ResultControllerSpec.php
@@ -11,6 +11,7 @@ use PhpSpec\ObjectBehavior;
 use PhpSpec\Wrapper\Collaborator;
 use Prophecy\Argument;
 use Psr\Container\ContainerInterface;
+use Psr\Log\LoggerInterface;
 use Shopware\Core\Content\Seo\SeoUrlPlaceholderHandlerInterface;
 use Shopware\Core\Framework\Adapter\Twig\TemplateFinder;
 use Shopware\Core\Framework\Event\NestedEventDispatcher;
@@ -35,6 +36,7 @@ class ResultControllerSpec extends ObjectBehavior
     private Collaborator $salesChannelContext;
     private Collaborator $twig;
     private Collaborator $seoUrlPlaceholderHandler;
+    private Collaborator $factfinderLogger;
 
     public function let(
         Request $request,
@@ -48,7 +50,8 @@ class ResultControllerSpec extends ObjectBehavior
         NestedEventDispatcher $nestedEventDispatcher,
         SystemConfigService $systemConfigService,
         Environment $twig,
-        SeoUrlPlaceholderHandlerInterface $seoUrlPlaceholderHandler
+        SeoUrlPlaceholderHandlerInterface $seoUrlPlaceholderHandler,
+        LoggerInterface $factfinderLogger,
     ): void {
         $this->request                  = $request;
         $this->config                   = $config;
@@ -57,7 +60,8 @@ class ResultControllerSpec extends ObjectBehavior
         $this->salesChannelContext      = $salesChannelContext;
         $this->twig                     = $twig;
         $this->seoUrlPlaceholderHandler = $seoUrlPlaceholderHandler;
-        $this->beConstructedWith($config, $pageLoader);
+        $this->factfinderLogger         = $factfinderLogger;
+        $this->beConstructedWith($config, $pageLoader, $factfinderLogger);
         $requestStack->getCurrentRequest()->willReturn($request);
         $container->get('request_stack')->willReturn($requestStack);
         $this->request->attributes = $attributes;